IntVarLike interface implementation (SmallInt)
When an integer constraint variable gets determined becomes an integer. For this reason integers can be seen as (compatible with) integer constraint variables (for instance to post constraints). For performance reasons gecode only accepts a subset of the values in the int type as valid domain elements. The consequence  of this is that the SmallInt type in the vm can represent a wider set of values and some of them not valid as domain elements.

The IntVarLike interface specifies a set of operations that must be implemented by normal constraint variables and by those small integers that are suitable as domain elements.